fix and clear docker images

+++ saldos/service/controller/index.js
@@ -3,12 +3,19 @@ const Saldo = require('./../models/Saldo');
 const fs = require('fs');
 const path = require('path');
 
-exports.getImporte = async ( req, res, next ) => {
-    const {placa} = req.params;
+exports.getAll = async ( req, res, next ) => {
     try {
-        const saldo = await Saldo.find({placa});
-        const importe = Number(parseInt(saldo.tiempo) * saldo.tarifa).toFixed(2);
-        res.status(200).send({importe,placa});        
+        const saldos = await Saldo.find({});
+        res.status(200).send(saldos);        
+    } catch (error) {
+        next(error);  
+    }
+}
+
+exports.deleteAll = async ( req, res, next ) => {
+    try {
+        const saldos = await Saldo.deleteMany({});
+        res.status(200).send(saldos);        
     } catch (error) {
         next(error);  
     }
@@ -19,24 +26,26 @@ exports.addTime = async ( req, res, next ) => {
     let { tiempo } = req.body;
     tiempo = tiempo || 0;
     try {
-        const updateSaldo = await Saldo.findOneAndUpdate({ placa },{$inc:{tiempo}});
+        const updateSaldo = await Saldo.findOneAndUpdate({ placa },{$inc:{tiempo}},{new:true});
+        
         let type, tarifa;
         /** obtener type */
         type = await repository.getTypeByPlaca(placa);
-
+        
+        if (!type ) throw new Error('placa no encontrada');
         /** obtener tarifa */
         tarifa = await repository.getTarifaByType(type);
 
         let result={},status = 200;
-        if (!updateSaldo || type !== 'oficial') {
+        if (!updateSaldo && type !== 'oficial') {
             const newSaldo = new Saldo({placa,tiempo,tarifa,type});
-            //res.status(201).send(newSaldo);
+            const save = await newSaldo.save();
             result = newSaldo;
             status = 201;
         }
         if (type === 'no residencial' && tiempo > 0) {
             const importe = Number(parseInt(tiempo) * tarifa).toFixed(2);
-            result= {placa,type,tarifa,tiempo,importe};    
+            result = {placa,type,tarifa,tiempo,importe};    
         }
         res.status(status).send(result);
     } catch (error) {
@@ -48,9 +57,13 @@ exports.monthBegins = async ( req, res, next ) => {
     try {
         const placasOficiales = await Saldo.find({ type: "oficial" });
         /** delete estancias oficiales */
-        await repository.deleteOficialEstancia(placasOficiales);
-        /** delete saldos oficiales */
-        await Saldo.deleteMany({ pacla: { $in: lacasOficiales }});
+        console.log(placasOficiales);
+        if (Array.isArray(placasOficiales) && placasOficiales.length) {
+            
+            await repository.deleteOficialEstancia(placasOficiales);
+            /** delete saldos oficiales */
+            await Saldo.deleteMany({ placa: { $in: placasOficiales }});
+        }
         /** reseteo tiempos a los residenciales */
         await Saldo.updateMany({ type: "residencial"}, { $set: { tiempo: 0 }});
 
@@ -64,13 +77,15 @@ exports.residentialPayment = async ( req, res, next ) => {
     const { filename } = req.body;
     try {
         let saldos = await Saldo.find({type: 'residencial'})
-        .map( saldo =>{
+        
+        saldos.map( saldo =>{
             saldo.importe = Number(parseInt(saldo.tiempo) * saldo.tarifa).toFixed(2);
         });
-        const uri = path.join(__dirname,'./../paymentsFiles',filename,'.txt');
+        
+        const uri = path.join(__dirname,'./../paymentsFiles',filename+'.txt');
         const titles = 'Núm. placa           Tiempo estacionado (min.)            Cantidad a pagar\n';
-
-        fs.appendFile(uri,titles,{ encoding: 'utf8' }, (err) =>{ if( err ) throw err });
+        console.log(uri);
+        fs.appendFile(uri,titles,{ encoding: 'utf8' }, (err) =>{ if( err ) console.log(err) });    
         for (let i = 0; i < saldos.length; i++) {
             const {placa, tiempo, importe} = saldos[i];
             const text= `${placa}                         ${tiempo}                     ${importe}\n`;
@@ -86,7 +101,7 @@ exports.residentialPayment = async ( req, res, next ) => {
 
         let readStream = fs.createReadStream(uri);
         readStream.pipe(res);
-
+       res.send(saldos);
     } catch (error) {
         next(error);  
     }
